+++ src/lib/pubkey/ed25519/ed25519_internal.h
@@ -15,11 +15,11 @@
 
 namespace Botan {
 
-inline uint64_t load_3(const uint8_t in[3]) {
-   return static_cast<uint64_t>(in[0]) | (static_cast<uint64_t>(in[1]) << 8) | (static_cast<uint64_t>(in[2]) << 16);
+inline uint32_t load_3(const uint8_t in[3]) {
+   return static_cast<uint32_t>(in[0]) | (static_cast<uint32_t>(in[1]) << 8) | (static_cast<uint32_t>(in[2]) << 16);
 }
 
-inline uint64_t load_4(const uint8_t* in) {
+inline uint32_t load_4(const uint8_t* in) {
    return load_le<uint32_t>(in, 0);
 }
